[DSPlayer] merged from trunk revs r30284:30383 fixed: don't use *= operator on an uninitialized variable
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/dsplayer@30384 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ project/BuildDependencies/DownloadBuildDeps.bat
@@ -1,14 +1,24 @@
 @ECHO OFF
 
+SETLOCAL
+
 SET CUR_PATH=%CD%
 SET XBMC_PATH=%CD%\..\..
-SET WGET=..\..\bin\wget
-SET ZIP=..\..\..\Win32BuildSetup\tools\7z\7za
+SET DL_PATH=%CD%\downloads
+SET TMP_PATH=%CD%\scripts\tmp
 
-rmdir lib /S /Q
-rmdir include /S /Q
+SET WGET=%CUR_PATH%\bin\wget
+SET ZIP=%CUR_PATH%\..\Win32BuildSetup\tools\7z\7za
+
+IF NOT EXIST %DL_PATH% md %DL_PATH%
+
+rem can't run rmdir and md back to back. access denied error otherwise.
+IF EXIST lib rmdir lib /S /Q
+IF EXIST include rmdir include /S /Q
+IF EXIST %TMP_PATH% rmdir %TMP_PATH% /S /Q
 md lib
 md include
+md %TMP_PATH%
 
 cd scripts
 
@@ -18,3 +28,5 @@ FOR /F "tokens=*" %%S IN ('dir /B "*_d.bat"') DO (
 )
 
 cd %CUR_PATH%
+
+rmdir %TMP_PATH% /S /Q
